+++ homeworks/13_splines/qspline.cc
@@ -8,7 +8,7 @@ namespace pp {
 
 qspline::qspline(const vector& x_in,
                  const vector& y_in)
-        : n(x_in.size()),
+        : n(static_cast<int>(x_in.size())),
           x(x_in),
           y(y_in),
           b(n-1),
